+++ src/components/manage/Widgets/ObjectBrowserWidget.jsx
@@ -348,7 +348,7 @@ export class ObjectBrowserWidgetComponent extends Component {
           >
             {items.map((item) => this.renderLabel(item))}
 
-            {items.length === 0 && this.props.mode === 'multiple' && (
+            {items.length === 0 && (
               <div className="placeholder" ref={this.placeholderRef}>
                 {this.props.placeholder ??
                   this.props.intl.formatMessage(messages.placeholder)}