+++ src/components/manage/Widgets/QueryWidget.jsx
@@ -7,11 +7,12 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { compose } from 'redux';
 import { connect } from 'react-redux';
+import { withRouter } from 'react-router';
 import { Button, Form, Grid, Input, Label } from 'semantic-ui-react';
 import { filter, remove, toPairs, groupBy, isEmpty, map } from 'lodash';
 import { defineMessages, injectIntl } from 'react-intl';
-import { getQuerystring } from '@plone/volto/actions';
-import { Icon } from '@plone/volto/components';
+import { getQuerystring, getQueryStringResults } from '@plone/volto/actions';
+import { Icon, ObjectBrowserWidget } from '@plone/volto/components';
 import { injectLazyLibs } from '@plone/volto/helpers/Loadable/Loadable';
 import cx from 'classnames';
 import config from '@plone/volto/registry';
@@ -38,6 +39,18 @@ const messages = defineMessages({
     id: 'querystring-widget-select',
     defaultMessage: 'Select…',
   },
+  no_items_selected: {
+    id: 'No items selected',
+    defaultMessage: 'No items selected',
+  },
+  currentPath: {
+    id: 'query-widget-currentPath',
+    defaultMessage: 'Current path (./)',
+  },
+  parentPath: {
+    id: 'query-widget-parentPath',
+    defaultMessage: 'Parent path (../)',
+  },
 });
 
 /**
@@ -61,6 +74,7 @@ export class QuerystringWidgetComponent extends Component {
     onEdit: PropTypes.func,
     onDelete: PropTypes.func,
     getQuerystring: PropTypes.func.isRequired,
+    getQueryStringResults: PropTypes.func.isRequired,
   };
 
   /**
@@ -92,6 +106,7 @@ export class QuerystringWidgetComponent extends Component {
     };
     this.onChangeValue = this.onChangeValue.bind(this);
     this.getWidget = this.getWidget.bind(this);
+    this.loadReferenceWidgetItem = this.loadReferenceWidgetItem.bind(this);
   }
 
   /**
@@ -106,6 +121,25 @@ export class QuerystringWidgetComponent extends Component {
     this.props.getQuerystring();
   }
 
+  loadReferenceWidgetItem(v) {
+    const loading = this.props.reference_request?.loading ?? false;
+    if (!loading && v?.length > 0) {
+      this.props.getQueryStringResults(
+        '/',
+        {
+          b_size: 1,
+          query: [
+            {
+              i: 'path',
+              o: 'plone.app.querystring.operation.string.absolutePath',
+              v: v + '::0',
+            },
+          ],
+        },
+        this.props.block + '_query_reference',
+      );
+    }
+  }
   /**
    * Get correct widget
    * @method getWidget
@@ -113,7 +147,7 @@ export class QuerystringWidgetComponent extends Component {
    * @param {number} index Row index.
    * @returns {Object} Widget.
    */
-  getWidget(row, index, Select) {
+  getWidget(row, index, Select, intl) {
     const props = {
       fluid: true,
       value: row.v,
@@ -121,7 +155,9 @@ export class QuerystringWidgetComponent extends Component {
     };
     const values = this.props.indexes[row.i].values;
 
-    switch (this.props.indexes[row.i].operators[row.o].widget) {
+    const operator = this.props.indexes[row.i].operators[row.o];
+
+    switch (operator.widget) {
       case null:
         return <span />;
       case 'DateWidget':
@@ -194,6 +230,64 @@ export class QuerystringWidgetComponent extends Component {
             />
           </Form.Field>
         );
+      case 'ReferenceWidget':
+        if (!this.props.reference) {
+          this.loadReferenceWidgetItem(props.value);
+        }
+
+        return (
+          <Form.Field style={{ flex: '1 0 auto', maxWidth: '92%' }}>
+            <ObjectBrowserWidget
+              style={{ flex: '1 0 auto' }}
+              onChange={(id, data) => {
+                const itemSelected = data.length > 0 ? data[0] : {};
+
+                this.onChangeValue(index, itemSelected.UID ?? '');
+                this.loadReferenceWidgetItem(itemSelected.UID);
+              }}
+              placeholder={intl.formatMessage(messages.no_items_selected)}
+              value={
+                props.value && this.props.reference
+                  ? [this.props.reference]
+                  : []
+              }
+              mode="link"
+              wrapped={false}
+            />
+          </Form.Field>
+        );
+      case 'RelativePathWidget':
+        const relativePathOptions = [
+          {
+            label: intl.formatMessage(messages.currentPath),
+            value: './',
+          },
+          {
+            label: intl.formatMessage(messages.parentPath),
+            value: '../',
+          },
+        ];
+        return (
+          <Form.Field style={{ flex: '1 0 auto', maxWidth: '92%' }}>
+            <Select
+              {...props}
+              className="react-select-container"
+              classNamePrefix="react-select"
+              options={relativePathOptions}
+              styles={customSelectStyles}
+              placeholder={this.props.intl.formatMessage(messages.select)}
+              theme={selectTheme}
+              components={{ DropdownIndicator, Option }}
+              onChange={(data) => {
+                this.onChangeValue(index, data.value);
+              }}
+              isMulti={false}
+              value={
+                relativePathOptions.filter((p) => p.value === props.value)?.[0]
+              }
+            />
+          </Form.Field>
+        );
       case 'autocomplete':
         const AutoCompleteComponent = config.widgets.widget.autocomplete;
         const vocabulary = { '@id': this.props.indexes[row.i].vocabulary };
@@ -211,14 +305,14 @@ export class QuerystringWidgetComponent extends Component {
             />
           </Form.Field>
         );
-      case 'ReferenceWidget':
+
       default:
         // if (row.o === 'plone.app.querystring.operation.string.relativePath') {
         //   props.onChange = data => this.onChangeValue(index, data.target.value);
         // }
         return (
           <Form.Field style={{ flex: '1 0 auto' }}>
-            <Input {...props} />
+            <Input {...props} description={operator.description} />
           </Form.Field>
         );
     }
@@ -403,7 +497,7 @@ export class QuerystringWidgetComponent extends Component {
                         </Button>
                       )}
                     </div>
-                    {this.getWidget(row, index, Select)}
+                    {this.getWidget(row, index, Select, intl)}
                     {this.props.indexes[row.i].operators[row.o].widget && (
                       <Button
                         onClick={(event) => {
@@ -495,10 +589,18 @@ export class QuerystringWidgetComponent extends Component {
 export default compose(
   injectIntl,
   injectLazyLibs(['reactSelect']),
+  withRouter,
   connect(
-    (state) => ({
-      indexes: state.querystring.indexes,
-    }),
-    { getQuerystring },
+    (state, props) => {
+      return {
+        indexes: state.querystring.indexes,
+        reference_request:
+          state.querystringsearch.subrequests[props.block + '_query_reference'],
+        reference:
+          state.querystringsearch.subrequests[props.block + '_query_reference']
+            ?.items?.[0],
+      };
+    },
+    { getQuerystring, getQueryStringResults },
   ),
 )(QuerystringWidgetComponent);