+++ CHANGELOG.md
@@ -3005,6 +3005,7 @@ Use next release instead: https://github.com/plone/volto/releases/tag/16.0.0-alp
 
 - Fixes in search block. Disable default live search. Added clear button for search input. Fixed facet dropdown clear button. Removed sort on label customization option. Layout improvements, CSS polishments. @kreafox @tiberiuichim
 - added default placeholder for videos to embed them more lightly @giuliaghisini
+- added widgets for path types in QueryWidget. @giuliaghisini
 - Added default placeholder for videos to embed them more lightly @giuliaghisini
 - Completed Romanian translation @sboghy